requirements.txt
# ============================================================
# ISOT Fake News Detection — Python Dependencies
# ============================================================

# Kaggle dataset download
kaggle>=1.6.14

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment variable management
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Machine learning / dataset splitting
scikit-learn>=1.3.0

# Deep learning framework (PyTorch)
torch>=2.0.0

# (Optional) NLP utilities — reserved for potential future use
# nltk>=3.8.0